fix: move logic to getFeedResponse.service.ts,  add error handlet to the empty catch, used @fastify/sensible
