+++ src/modules/feedParser/routes/feedParser.route.ts
@@ -1,9 +1,9 @@
 import type { JsonSchemaToTsProvider } from "@fastify/type-provider-json-schema-to-ts";
 import type { FastifyInstance } from "fastify";
 import fp from "fastify-plugin";
-import { getFeedCacheByUrl, getLatestItems } from "../../../services/dbService";
 import { schema } from "../schemas/getFeedData.schema";
-import { isUpstream4xx, parseFeed } from "../services/feedParserService";
+import { isUpstream4xx } from "../services/feedParserService";
+import { getFeedResponse } from "../services/getFeedResponse.service";
 
 export default fp(async function routes(app: FastifyInstance) {
   const r = app.withTypeProvider<JsonSchemaToTsProvider>();
@@ -11,48 +11,16 @@ export default fp(async function routes(app: FastifyInstance) {
   r.get("/feed", { schema }, async (req, reply) => {
     try {
       const q = req.query as { url?: string; force?: "0" | "1" };
-
       const url = q.url ?? app.config.DEFAULT_FEED_URL;
       if (!url) return reply.badRequest("url is required or DEFAULT_FEED_URL must be set");
-      const _force = q.force === "1";
 
-      if (!_force) {
-        const cache = await getFeedCacheByUrl(app.prisma, url);
-        if (cache) {
-          const cachedItems = await getLatestItems(app.prisma, url, 50);
-          return reply.send({
-            sourceUrl: url,
-            title: cache.title ?? "",
-            itemsCount: cachedItems.length,
-            items: cachedItems.map((i) => ({
-              guid: i.guid ?? "",
-              title: i.title ?? "",
-              link: i.link ?? "",
-              isoDate: i.isoDate ? i.isoDate.toISOString() : null,
-            })),
-          });
-        }
-      }
-
-      const { title, items } = await parseFeed(url);
-      const responseItems = items.map((i) => ({
-        guid: i.guid,
-        title: i.title,
-        link: i.link,
-        isoDate: i.isoDate ? i.isoDate.toISOString() : null,
-      }));
-      return reply.send({
-        sourceUrl: url,
-        title,
-        itemsCount: responseItems.length,
-        items: responseItems,
-      });
+      const data = await getFeedResponse(app, url, q.force === "1");
+      return reply.send(data);
     } catch (err) {
+      req.log.error({ err, query: req.query }, "feed route failed");
       const e = err as { code?: string; message?: string };
 
-      if (isUpstream4xx(err)) {
-        return reply.badRequest(e.message || "Upstream 4xx");
-      }
+      if (isUpstream4xx(err)) return reply.badRequest(e.message || "Upstream 4xx");
       if (e.code === "ENOTFOUND" || e.code === "ECONNREFUSED" || e.code === "ECONNRESET") {
         return reply.badGateway(e.message || "Upstream network error");
       }
